refactor: update WalletPrompt import path and add wallet check in OrderDetails
- Changed the import path for WalletPrompt in portfolio page to reflect new location.
- Removed WalletPrompt component from its previous location.
- Added a wallet connection check in OrderDetails component to prompt users to connect their wallet if not already connected.
