+++ app/order-book-dex/src/app/portfolio/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { WalletPrompt } from "@/components/portfolio/wallet-prompt";
+import { WalletPrompt } from "@/components/wallet-prompt";
 import { useWallet } from "@solana/wallet-adapter-react";
 import { PortfolioHeader } from "./header";
 import { PortfolioOpenOrders } from "./open-orders";
